opti: refactor code
