+++ packages/components/connectWallet/ConnectAdenaButton.tsx
@@ -6,12 +6,9 @@ import { useFeedbacks } from "../../context/FeedbacksProvider";
 
 import adenaSVG from "@/assets/icons/adena.svg";
 import { useAdenaStore } from "@/context/WalletsProvider/adena/useAdenaStore";
+import { useAdenaUtils } from "@/context/WalletsProvider/adena/useAdenaUtils";
 import { useSelectedNetworkInfo } from "@/hooks/useSelectedNetwork";
-import {
-  getGnoNetworkFromChainId,
-  NetworkInfo,
-  NetworkKind,
-} from "@/networks/index";
+import { getGnoNetworkFromChainId } from "@/networks/index";
 import { setIsAdenaConnected } from "@/store/slices/settings";
 import { useAppDispatch } from "@/store/store";
 
@@ -23,51 +20,7 @@ export const ConnectAdenaButton: React.FC<{
   const { state, setState } = useAdenaStore();
   const selectedNetworkInfo = useSelectedNetworkInfo();
 
-  const switchAdenaNetwork = async (
-    adena: any,
-    selectedNetworkInfo: NetworkInfo | undefined,
-  ) => {
-    if (!adena) return;
-    if (selectedNetworkInfo?.kind !== NetworkKind.Gno) return;
-
-    try {
-      const res = await adena.SwitchNetwork(selectedNetworkInfo?.chainId);
-
-      if (res.status === "success") {
-        setState({ chainId: res.data.chainId });
-        return;
-      }
-
-      console.warn(res);
-
-      if (res.type === "UNADDED_NETWORK") {
-        const res = await adena.AddNetwork({
-          chainId: selectedNetworkInfo.chainId,
-          rpcUrl: selectedNetworkInfo.endpoint,
-          chainName: selectedNetworkInfo.displayName,
-        });
-
-        if (res.status === "failure") {
-          throw Error(res.message);
-        }
-
-        await switchAdenaNetwork(adena, selectedNetworkInfo);
-        return;
-      }
-
-      throw Error(res.message);
-    } catch (err) {
-      console.error(err);
-      if (err instanceof Error) {
-        setToast({
-          type: "error",
-          message: err.message,
-          mode: "normal",
-          title: "Failed to connect to Adena (2)",
-        });
-      }
-    }
-  };
+  const { switchAdenaNetwork } = useAdenaUtils();
 
   const handlePress = async () => {
     try {
