+++ packages/context/WalletsProvider/adena/index.ts
@@ -2,18 +2,13 @@ import { useCallback, useEffect, useMemo, useState } from "react";
 import { useSelector } from "react-redux";
 
 import { useAdenaStore } from "./useAdenaStore";
+import { useAdenaUtils } from "./useAdenaUtils";
 import { useAppDispatch } from "../../../store/store";
 import { Wallet } from "../wallet";
 
 import { useFeedbacks } from "@/context/FeedbacksProvider";
 import { useSelectedNetworkInfo } from "@/hooks/useSelectedNetwork";
-import {
-  GnoNetworkInfo,
-  NetworkInfo,
-  NetworkKind,
-  allNetworks,
-  getUserId,
-} from "@/networks";
+import { NetworkKind, allNetworks, getUserId } from "@/networks";
 import {
   selectIsAdenaConnected,
   setIsAdenaConnected,
@@ -36,6 +31,8 @@ export const useAdena: () => UseAdenaResult = () => {
 
   const { state, setState } = useAdenaStore();
   const [ready, setReady] = useState(false);
+
+  const { addAdenaNetwork, switchAdenaNetwork } = useAdenaUtils();
 
   const fetchAccount = useCallback(
     async (
@@ -80,70 +77,6 @@ export const useAdena: () => UseAdenaResult = () => {
       setReady(true);
     },
     [selectedNetworkInfo, setState],
-  );
-
-  const addAdenaNetwork = useCallback(
-    async (adena: any, selectedNetworkInfo: GnoNetworkInfo) => {
-      const res = await adena.AddNetwork({
-        chainId: selectedNetworkInfo.chainId,
-        rpcUrl: selectedNetworkInfo.endpoint,
-        chainName: selectedNetworkInfo.displayName,
-      });
-
-      if (res.status === "failure") {
-        throw Error(res.message);
-      }
-    },
-    [],
-  );
-
-  const switchAdenaNetwork = useCallback(
-    async (adena: any, selectedNetworkInfo: NetworkInfo | undefined) => {
-      if (!adena) return;
-      if (selectedNetworkInfo?.kind !== NetworkKind.Gno) return;
-
-      try {
-        const network = await adena.GetNetwork();
-        if (network.status === "failure") {
-          if (network.type === "NOT_CONNECTED") return;
-          throw Error(network.message);
-        }
-
-        const adenaChainId = network.data.chainId;
-
-        if (adenaChainId === selectedNetworkInfo?.chainId) {
-          return;
-        }
-
-        const res = await adena.SwitchNetwork(selectedNetworkInfo?.chainId);
-
-        if (res.status === "success") {
-          setState({ chainId: res.data.chainId });
-          return;
-        }
-
-        console.warn(res);
-
-        if (res.type === "UNADDED_NETWORK") {
-          await addAdenaNetwork(adena, selectedNetworkInfo);
-          await switchAdenaNetwork(adena, selectedNetworkInfo);
-          return;
-        }
-
-        throw Error(res.message);
-      } catch (err) {
-        console.error(err);
-        if (err instanceof Error) {
-          setToast({
-            type: "error",
-            message: err.message,
-            mode: "normal",
-            title: "Failed to connect to Adena (3)",
-          });
-        }
-      }
-    },
-    [setToast, addAdenaNetwork, setState],
   );
 
   useEffect(() => {
